+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.7.0
+- Added support to query IS NULL and IS NOT NULL in query builder comparator.
+- Added method to get count from a query builder.
+
 ## v0.6.0
 - Added support to multiple order by in query.